Feature: Refactored statistic operations into a statistics directory
