+++ src/statistics/mean.py
