+++ src/statistics/mode.py
