+++ src/statistics/variance.py
